Accurate sigop/sighash accounting and limits (INCOMPLETE 0.12 version)
Adds a ValidationCostTracker class that is passed to
CheckInputs() / CScriptCheck() to keep track of the exact number
of signature operations required to validate a block, and the
exact number of bytes hashed to compute signature hashes.

Also extends CHashWriter to keep track of number of bytes hashed.

Signature operations per block are limited to MAX_BLOCK_SIGOPS
(unchanged at 20,000)

Bytes hashed to compute signatures is limited to MAX_BLOCK_SIGHASH
(1.3 GB in this commit).

--

This version includes significant differences written by jtoomim
compared to the commit made for 0.11.2 by Gavin due to the massive
changes made to CreateNewBlock in Core 0.12. Since transactions are
no longer validated inside CreateNewBlock in 0.12, the bytes hashed
and sigops for each transaction cannot be calculated inside
CreateNewBlock. Instead, they must be precomputed and stored inside
the CTxMempoolEntry object. This commit is a partial implementation of
that process. It records the sigops and sighash when transactions enter
the mempool through AcceptToMemoryPool. However, it does not yet
do so for transactions that enter via reorgs or via the wallet.

In addition, the changes made in this code breaks the
miner_tests.cpp stuff, since that code expects an exception to
be raised, but this code will simply refuse to include transactions
that would exceed the sigops/sighash limit. This behavior is okay,
but it does mean that the unit tests need to be rewritten to
check for whether the naughty transaction was included in the block
or not.

Conflicts:
	src/main.cpp
	src/miner.cpp
	src/script/interpreter.h